Enum twilight_validate::command::CommandValidationErrorType

#[non_exhaustive]
pub enum CommandValidationErrorType {

    CountInvalid,
    CommandTooLarge {
        characters: usize,
    },
    DescriptionInvalid,
    NameLengthInvalid,
    NameCharacterInvalid {
        character: char,
    },
    OptionDescriptionInvalid,
    OptionNameLengthInvalid,
    OptionNameNotUnique {
        option_index: usize,
    },
    OptionNameCharacterInvalid {
        character: char,
    },
    OptionChoiceNameLengthInvalid,
    OptionChoiceStringValueLengthInvalid,
    OptionsCountInvalid,
    OptionsRequiredFirst {
        index: usize,
    },
    PermissionsCountInvalid,

}

Type of CommandValidationError that occurred.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CountInvalid

Too many commands have been provided.

The maximum number of commands is defined by GUILD_COMMAND_LIMIT.

CommandTooLarge

Fields

characters: usize

Provided number of codepoints.

Combined values of the command are larger than COMMAND_TOTAL_LENGTH.

This includes name or the longest name localization, description or the longest description localization of the command and its options and the choice names or the longest name localization and the choice value if it is a string choice.

DescriptionInvalid

Command description is invalid.

NameLengthInvalid

Command name length is invalid.

NameCharacterInvalid

Fields

character: char

Invalid character.

Command name contain an invalid character.

OptionDescriptionInvalid

Command option description is invalid.

OptionNameLengthInvalid

Command option name length is invalid.

OptionNameNotUnique

Fields

option_index: usize

Index of the option that has a duplicated name.

Command option name is non-unique.

OptionNameCharacterInvalid

Fields

character: char

Invalid character.

Command option name contain an invalid character.

OptionChoiceNameLengthInvalid

Command option choice name length is invalid.

OptionChoiceStringValueLengthInvalid

String command option choice value length is invalid.

OptionsCountInvalid

Command options count invalid.

OptionsRequiredFirst

Fields

index: usize

Index of the option that failed validation.

Required command options have to be passed before optional ones.

PermissionsCountInvalid

More than 10 permission overwrites were set.