Function twilight_validate::component::action_row

pub fn action_row(
    action_row: &ActionRow,
) -> Result<(), ComponentValidationError>

Ensure that an action row is correct.

Errors

Returns an error of type ActionRowComponentCount if the action row has too many components in it.

Returns an error of type InvalidChildComponent if the provided nested component is an ActionRow. Action rows can not contain another action row.

Refer to button for potential errors when validating a button in the action row.

Refer to select_menu for potential errors when validating a select menu in the action row.

Refer to text_input for potential errors when validating a text input in the action row.