Module twilight_validate::sticker

Constants, error types, and functions for validating Sticker fields.

Structs

StickerValidationError

Error created if validation of a sticker field fails.

Enums

StickerValidationErrorType

Type of StickerValidationError that occurred.

Constants

STICKER_DESCRIPTION_LENGTH_MAX

Maximum length of a sticker description.

STICKER_DESCRIPTION_LENGTH_MIN

Minimum length of a sticker description.

STICKER_NAME_LENGTH_MAX

Maximum length of a sticker name.

STICKER_NAME_LENGTH_MIN

Minimum length of a sticker name.

STICKER_TAGS_LENGTH_MAX

Maximum length of the sticker’s tags.

STICKER_TAGS_LENGTH_MIN

Minimum length of the sticker’s tags.

Functions

description

Ensure that a sticker’s description is correct.

name

Ensure that a sticker’s name is correct.

tags

Ensure that a sticker’s tags is correct.