Crate twilight_validate

twilight-validate

twilight-validate contains constants, methods, and error types for validating request parameters in the twilight-rs ecosystem.

This crate is used internally by twilight-http, but may also be used for manually validating any models from twilight-model.

Modules

channel

Constants, error types, and functions for validating channel fields.

command

Constants, error types, and functions for validating [Command]s.

component

Constants, error types, and functions for validating [Component]s.

embed

Constants, error types, and functions for validating [Embed]s.

message

Constants, error types, and functions for validating Message fields.

request

Contains all other input validation functions.

sticker

Constants, error types, and functions for validating Sticker fields.