Module twilight_validate::channel

Constants, error types, and functions for validating channel fields.

Structs

ChannelValidationError

Returned when the channel can not be updated as configured.

Enums

ChannelValidationErrorType

Type of ChannelValidationError that occurred.

Constants

CHANNEL_BITRATE_MIN

Minimum bitrate of a voice channel.

CHANNEL_FORUM_TOPIC_LENGTH_MAX

Maximum length of a forum channel’s topic.

CHANNEL_NAME_LENGTH_MAX

Maximum length of a channel’s name.

CHANNEL_NAME_LENGTH_MIN

Minimum length of a channel’s name.

CHANNEL_RATE_LIMIT_PER_USER_MAX

Maximum length of a channel’s rate limit per user.

CHANNEL_TOPIC_LENGTH_MAX

Maximum length of a channel’s topic.

Functions

bitrate

Ensure a channel’s bitrate is collect.

forum_topic

Ensure a forum channel’s topic’s length is correct.

is_thread

Ensure a channel is a thread.

name

Ensure a channel’s name’s length is correct.

rate_limit_per_user

Ensure a channel’s rate limit per user is correct.

topic

Ensure a channel’s topic’s length is correct.