Module twilight_validate::command

Constants, error types, and functions for validating Commands.

Structs

CommandValidationError

Error created when a Command is invalid.

Enums

CommandValidationErrorType

Type of CommandValidationError that occurred.

Constants

CHOICES_LIMIT

Maximum number of choices an option can have.

DESCRIPTION_LENGTH_MAX

Maximum length of a command’s description.

DESCRIPTION_LENGTH_MIN

Minimum length of a command’s description.

GUILD_COMMAND_LIMIT

Maximum number of commands an application may have in an individual guild.

GUILD_COMMAND_PERMISSION_LIMIT

Maximum number of permission overwrites an application may have in an individual guild command.

NAME_LENGTH_MAX

Maximum length of a command’s name.

NAME_LENGTH_MIN

Minimum length of a command’s name.

OPTIONS_LIMIT

Maximum amount of options a command may have.

OPTION_DESCRIPTION_LENGTH_MAX

Maximum length of a command’s description.

OPTION_DESCRIPTION_LENGTH_MIN

Minimum length of a command’s description.

OPTION_NAME_LENGTH_MAX

Maximum length of a command’s name.

OPTION_NAME_LENGTH_MIN

Minimum length of a command’s name.

Functions

chat_input_name

Validate the name of a ChatInput command.

command

Validate a Command.

description

Validate the description of a Command.

guild_permissions

Validate the number of guild command permission overwrites.

name

Validate the name of a User or Message command.

option

Validate a single CommandOption.

option_name

Validate the name of a CommandOption.

options

Validate a list of command options for count, order, and internal validity.