Module twilight_validate::embed

Constants, error types, and functions for validating [Embed]s.

Structs

EmbedValidationError

An embed is not valid.

Enums

EmbedValidationErrorType

Type of EmbedValidationError that occurred.

Constants

AUTHOR_NAME_LENGTH

The maximum embed author name length in codepoints.

COLOR_MAXIMUM

The maximum accepted color value.

DESCRIPTION_LENGTH

The maximum embed description length in codepoints.

EMBED_TOTAL_LENGTH

The maximum combined embed length in codepoints.

FIELD_COUNT

The maximum number of fields in an embed.

FIELD_NAME_LENGTH

The maximum length of an embed field name in codepoints.

FIELD_VALUE_LENGTH

The maximum length of an embed field value in codepoints.

FOOTER_TEXT_LENGTH

The maximum embed footer length in codepoints.

TITLE_LENGTH

The maximum embed title length in codepoints.

Functions

chars

Calculate the total character count of an embed.

embed

Ensure an embed is correct.