Module twilight_validate::message

Constants, error types, and functions for validating Message fields.

Structs

MessageValidationError

A message is not valid.

Enums

MessageValidationErrorType

Type of MessageValidationError that occurred.

Constants

EMBED_COUNT_LIMIT

Maximum number of embeds that a message may have.

MESSAGE_CONTENT_LENGTH_MAX

Maximum length of message content.

STICKER_MAX

Maximum amount of stickers.

Functions

components

Ensure a list of components is correct.

content

Ensure a message’s content is correct.

embeds

Ensure a list of embeds is correct.

stickers

Ensure that the amount of stickers in a message is correct.